+++ test/central_web/controllers/admin/user_controller_test.exs
@@ -12,7 +12,7 @@ defmodule CentralWeb.Admin.UserControllerTest do
   end
 
   @create_attrs %{
-    colour: "some colour",
+    colour: "#AA0000",
     email: "some email",
     icon: "fa-regular fa-home",
     name: "some name",
@@ -22,7 +22,7 @@ defmodule CentralWeb.Admin.UserControllerTest do
     data: "{}"
   }
   @update_attrs %{
-    colour: "some updated colour",
+    colour: "#0000AA",
     icon: "fa-solid fa-wrench",
     name: "some updated name",
     permissions: [],
@@ -248,7 +248,7 @@ defmodule CentralWeb.Admin.UserControllerTest do
       # assert redirected_to(conn) == Routes.admin_user_path(conn, :index)
 
       conn = get(conn, Routes.admin_user_path(conn, :show, user))
-      assert html_response(conn, 200) =~ "some updated colour"
+      assert html_response(conn, 200) =~ "#0000AA"
     end
 
     test "renders errors when data is invalid", %{conn: conn, user: user} do